requirements.txt
django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
python-dotenv==1.0.0
psycopg2-binary==2.9.9
whitenoise==6.5.0
django-filter==23.3
celery==5.3.4
redis==5.0.1